+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Fixed
+
+- Use `String` type for `HAS.condition.column` in `@whereConditions` instead of incorrectly inheriting the parent's column enum type https://github.com/nuwave/lighthouse/pull/2725
+
 ## v6.64.1
 
 ### Fixed